// frontend/src/platforms/catalogue.generated.ts
export type CatalogueKind = 'command' | 'code' | 'download' | 'info'

export interface CatalogueType {
  name: string
  displayName?: string
  installation?: string // slug into installations
}

export interface CatalogueInstallation {
  name: string
  kind: CatalogueKind
  commandTemplate?: string
  qualifier?: string
  instructions?: string
  link?: string
  altLink?: string
  services?: { application: number; port?: number }[]
}